Extract DRL comment tests into separate test class
